refactor(screenscraper): promote scraping summary dialog to a widget
Extract the embedded private `_ScrapingSummaryDialogContent` StatefulWidget
out of `screenscraper_service.dart` into a public `ScrapingSummaryDialog`
in `lib/widgets/scraping_summary_dialog.dart`. Body moved verbatim — the
gamepad-layer lifecycle (initState/dispose), build tree, and the
`_buildCompactStatRow` helper are byte-identical; the only changes are the
public rename and a `super.key` constructor param.

`ScreenScraperService._showScrapingSummaryDialog` now builds the public
widget. Four imports used only by the moved dialog (material_symbols_icons,
game_service, gamepad_nav, flutter_screenutil) are dropped from the service.
Host 1734 -> 1538 lines. Zero behavior change.

Verified: dart format clean, flutter analyze No issues found project-wide,
207/207 tests pass.
